requirements.txt
python-dotenv==1.0.0
aiogram==2.25.1
gino==1.0.1
SQLAlchemy==1.3.24
asyncpg==0.27.0
requests==2.30.0
httplib2==0.22.0
oauth2client==4.1.3
google-api-python-client==2.87.0